move SkikoViewController
