+++ samples/SkiaMultiplatformSample/src/iosMain/kotlin/org/jetbrains/skiko/sample/getSkikoViewContoller.kt
@@ -16,3 +16,35 @@ fun getSkikoViewContoller(): UIViewController = SkikoViewController(
         }
     )
 )
+
+@ExportObjCClass
+class SkikoViewController : UIViewController {
+    @OverrideInit
+    constructor() : super(nibName = null, bundle = null)
+    @OverrideInit
+    constructor(coder: NSCoder) : super(coder)
+
+    constructor(skikoUIView: SkikoUIView) : this() {
+        this.skikoUIView = skikoUIView
+    }
+
+    private var skikoUIView: SkikoUIView? = null
+
+    override fun loadView() {
+        if (skikoUIView == null) {
+            super.loadView()
+        } else {
+            this.view = skikoUIView!!.load()
+        }
+    }
+
+    override fun viewDidLoad() {
+        super.viewDidLoad()
+        skikoUIView?.showScreenKeyboard()
+    }
+
+    // viewDidUnload() is deprecated and not called.
+    override fun viewDidDisappear(animated: Boolean) {
+        skikoUIView?.detach()
+    }
+}
